% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal.R
\name{all_equal}
\alias{all_equal}
\title{Test is all values in a vector are equal.}
\usage{
all_equal(x, na.rm = FALSE)
}
\arguments{
\item{x}{[atomic_vector]}

\item{na.rm}{[logical]  Defaults to FALSE}
}
\value{
\code{all_equal}: [Logical(1)]
}
\description{
Test is all values in a vector are equal.
}
\examples{
 all_equal(rep('a', 5))
 all_equal(c('b', rep('a', 5)))
 all_equal(c(NA_character_, rep('a', 5)))
 all_equal(c(NA_character_, rep('a', 5)), na.rm = TRUE)
}
