% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_error_message.R
\name{as_error_message}
\alias{as_error_message}
\title{Add class 'error_message' to a character string or vector.}
\usage{
as_error_message(x)
}
\arguments{
\item{x}{\code{[character]}  NULL is ok.}
}
\value{
\code{as_error_message}: \code{[character]} character with class error_message
}
\description{
Add class 'error_message' to a character string or vector.
}
\examples{
 as_error_message(x = 'Error message')
}
