% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter of a string.}
\usage{
capitalize(x)
}
\arguments{
\item{x}{\code{[character]}}
}
\value{
\code{capitalize}: [character]
}
\description{
Capitalize the first letter of a string.
}
\examples{
 capitalize(c('dog', 'hello, sir.', '1st time', 'ALREADY CAPS'))
}
