% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{check_length}
\alias{check_length}
\title{Check the length of an r object.}
\usage{
check_length(x, len, null.ok = FALSE)
}
\arguments{
\item{len}{\code{[integerish]}}

\item{null.ok}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}
}
\value{
\code{check_length}: \code{[TRUE|character]}
}
\description{
Check the length of an r object.
}
\examples{
 check_length(letters, len = 26, null.ok = FALSE)
 check_length(letters, len = 25, null.ok = FALSE)
 check_length(letters[1:5], len = c(0, 26), null.ok = FALSE)
 check_length(letters[0], len = c(0, 26), null.ok = FALSE)
 check_length(NULL, len = c(0, 26), null.ok = FALSE)
 check_length(NULL, len = c(0, 26), null.ok = TRUE)
}
