% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_empty.R
\name{drop_empty}
\alias{drop_empty}
\title{Drop Null or length 0 list arguments.}
\usage{
drop_empty(x)
}
\arguments{
\item{x}{\code{[list]}  NULL is ok.}
}
\value{
\code{drop_empty}: \code{[list]}
}
\description{
Drop Null or length 0 list arguments.
}
\examples{
 drop_empty(x = list(one = NULL, two = character(), three = 1))
}
