% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{error_message}
\alias{error_message}
\title{Set and collect error messages.  For use in tryCatch.}
\usage{
error_message(message)
}
\arguments{
\item{message}{the message to return if the execution in tryCatch throws an error}
}
\value{
a message of class error
}
\description{
Set and collect error messages.  For use in tryCatch.
}
\examples{
 message <- tryCatch(1+a_nonexistent_variable,error=error_message("Execution Error"))
 is_error(message)
 print(message)
}
