% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in.R
\name{\%all_in\%}
\alias{\%all_in\%}
\title{Convience function for testing if all values in vector 'a' are present in vector 'b'.}
\usage{
a \%all_in\% b
}
\arguments{
\item{a}{target vector}

\item{b}{key vector}
}
\value{
logical: true if all values in 'a' are also in 'b'.
}
\description{
Convience function for testing if all values in vector 'a' are present in vector 'b'.
}
\examples{
 a <- c("A", "A", "A", "B", "B", "C")
 b <- c("A", "B", "C")
 a \%all_in\% b
 a <- c("A", "A", "A", "B", "B", "C", "D" )
 b <- c("A", "B", "C")
 a \%all_in\% b
}
