% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep.R
\name{\%grep\%}
\alias{\%grep\%}
\title{Convience function for grep.}
\usage{
x \%grep\% pattern
}
\arguments{
\item{x}{a character vector}

\item{pattern}{a pattern to match.  if \code{length(pattern)>1}, pattern will be collapsed by '|' (or).}
}
\value{
a vector of matches
}
\description{
Convience function for grep.
}
\examples{
a <- c("catfish", "catty","doggy", "dog", "man", "boy")
b <- c("cat","dog")
a \%grep\% b
}
