% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IN.R
\name{\%IN\%}
\alias{\%IN\%}
\title{Convience function for subsetting values.}
\usage{
a \%IN\% b
}
\arguments{
\item{a}{target vector}

\item{b}{key vector}
}
\value{
unique values in 'a' that are  in 'b'
}
\description{
Convience function for subsetting values.
}
\examples{
 a <- c("A", "A", "A", "B", "B", "C")
 b <- c("A", "A", "C", "C")
 a \%IN\% b
}
