% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrep.R
\name{\%ngrep\%}
\alias{\%ngrep\%}
\title{Inverse grep}
\usage{
x \%ngrep\% pattern
}
\arguments{
\item{pattern}{vector of patterns to match. if \code{length(pattern)>1}, pattern will be collapsed by '|' (or).}

\item{a}{target vector}
}
\value{
a vector of matches
}
\description{

}
\examples{
 a <- c("wages.csv", "wages.txt", "wages.xlsx")
 pattern <- c(".txt", ".csv")
 a \%ngrep\% pattern
}
