% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NIN.R
\name{\%NIN\%}
\alias{\%NIN\%}
\title{Convience function for reverse subsetting values.}
\usage{
a \%NIN\% b
}
\arguments{
\item{a}{target vector}

\item{b}{key vector}
}
\value{
values in 'a' that are not in 'b'
}
\description{
Convience function for reverse subsetting values.
}
\examples{
 a <- c("A", "A", "A", "B", "B", "C")
 b <- c("A", "A", "C", "C")
 a \%NIN\% b
}
