% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{\%or\%}
\alias{\%or\%}
\title{Logical test for non-zero, non-null, and TRUE.}
\usage{
x \%or\% y
}
\arguments{
\item{x}{an object}

\item{y}{an object}
}
\value{
Returns 'x' if it is length > zero (including character count), non-null,not missing, or TRUE.  Otherwise returns 'y'
}
\description{
Logical test for non-zero, non-null, and TRUE.
}
\examples{
 x <- NULL
 y <- 2
 x \%or\% y
 x <- 4
 y <- 2
 x \%or\% y
}
