% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UIN.R
\name{\%UIN\%}
\alias{\%UIN\%}
\title{Convience function for subsetting unique values.}
\usage{
a \%UIN\% b
}
\arguments{
\item{a}{target vector}

\item{b}{key vector}
}
\value{
unique values in 'a' that are  in 'b'
}
\description{
Convience function for subsetting unique values.
}
\examples{
 a <- c("A", "A", "A", "B", "B", "C")
 b <- c("A", "A", "C", "C")
 a \%UNI\% b
}
