% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNIN.R
\name{\%UNIN\%}
\alias{\%UNIN\%}
\title{Convience function for reverse subsetting unique values.}
\usage{
a \%UNIN\% b
}
\arguments{
\item{a}{target vector}

\item{b}{key vector}
}
\value{
unique values in 'a' that are not in 'b'
}
\description{
Convience function for reverse subsetting unique values.
}
\examples{
 a <- c("A", "A", "A", "B", "B", "C")
 b <- c("A", "A", "C", "C")
 a \%UNIN\% b
}
