% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fn.R
\name{import_fn}
\alias{import_fn}
\title{Reconstruct an existing function in the current or specified environment}
\usage{
import_fn(fn, env = caller_env())
}
\arguments{
\item{fn}{existing function name}

\item{env}{an environment. Defaults to the current environment.}
}
\value{
the same function recreated in the current or specified environment
}
\description{
Reconstruct an existing function in the current or specified environment
}
