% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{Check is an arguement has no rows, is NULL, or zero length.}
\usage{
is_empty(x)
}
\value{
\code{is_empty}: \code{[logical(1)]}
}
\description{
Check is an arguement has no rows, is NULL, or zero length.
}
\examples{
 is_empty(x = NULL)
 is_empty(x = 1)
 is_empty(x = c(1:2))
 is_empty(x = data.frame())
 is_empty(x = data.frame(1))
 is_empty(x = list())
 is_empty(x = list(1))
 is_empty(x = 'a')
 is_empty(x = '')
}
