% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_order.R
\name{match_order}
\alias{match_order}
\title{Get the index order of vector x to match the order of a second vector}
\usage{
match_order(x, by)
}
\arguments{
\item{x}{a vector}

\item{by}{the look up vector}
}
\value{
the index order of vector 'x' if it were to match its order to the second vector
}
\description{
Get the index order of vector x to match the order of a second vector
}
\examples{
 x <- c("K","G","R","A","C")
 by <- c("A","R","C","K","G")
 I=match_order(x,by)
 I
 x[I]
 x[I]==by
}
