% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_precision.R
\name{n_precision}
\alias{n_precision}
\title{get the numeric precision of a vector.}
\usage{
n_precision(x)
}
\arguments{
\item{x}{\code{[numeric]}}
}
\value{
\code{n_precision}: \code{[numeric(length(x))]}
}
\description{
get the numeric precision of a vector.
}
\examples{
 n_precision(x = c(23.1, 23, 120, 234.021, NA))
}
