% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Build a file path.}
\usage{
path(...)
}
\arguments{
\item{...}{the pieces of the path}
}
\value{
\code{path}: character(file_path)
}
\description{
Build a file path.
}
\examples{
 path(getwd(), 'test', 'testtest.R')
}
