% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_vec.R
\name{split_vec}
\alias{split_vec}
\title{Split a vector in to two based on a logical test.}
\usage{
split_vec(x, y)
}
\arguments{
\item{x}{\code{[atomic_vector]}}

\item{y}{\code{[logical]}  Must have an exact length of \code{length(x)}.}
}
\value{
\code{split_vec}: list(x[TRUE],x[FALSE])
}
\description{
Split a vector in to two based on a logical test.
}
\examples{
 x <- c('app', 'apps', 'apple', 'bet', 'cat')
 split_vec(x, x \%detect\% 'app')
 x <- c(1:4)
 y <- x > 2
 split_vec(x, y)
}
