% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_escape.R
\name{str_escape}
\alias{str_escape}
\title{Escape a regex pattern.}
\usage{
str_escape(x = NULL)
}
\arguments{
\item{x}{\code{[character]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{str_escape}: \code{[character]} a character vector with length equal to x.
}
\description{
Escape a regex pattern.
}
\examples{
 str_escape(x = c('var(1)', '[sd]', 'test$list', 'a.t', '2^4'))
 str_escape(x = NULL)
}
