% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_modify.R
\name{str_modify}
\alias{str_modify}
\title{Modify parts of a string.}
\usage{
str_modify(x, pattern, func, .open = "&&{", .close = "}&&")
}
\arguments{
\item{x}{\code{[character]}}

\item{pattern}{\code{[string]}}

\item{func}{\code{[function]}}
}
\value{
\code{str_modify}: [character]
}
\description{
Modify parts of a string.
}
\examples{
 x = c('[character] this a character vector', '[numeric] this is a numeric vector')
 str_modify(x, pattern = '^?\\\\[.*\\\\]', func = function(x) paste0('\\\\code{',
 x, '}'))
}
