% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_modify_all.R
\name{str_modify_all}
\alias{str_modify_all}
\title{Modify a string.}
\usage{
str_modify_all(x, pattern, func)
}
\arguments{
\item{x}{\code{[character]}}

\item{pattern}{\code{[string]}}

\item{func}{\code{[function]}}
}
\value{
\code{str_modify_all}: [character]
}
\description{
Modify a string.
}
\examples{
 x = c('[characterORnumeric] this a [character] or [numeric] vector',
 '[numeric] this is a [numeric] vector')
 str_modify_all(x, pattern = '\\\\[\\\\w+\\\\]', func = function(x) paste0('\\\\code{',
 x, '}'))
}
