% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_parsing.R
\name{try_parsing}
\alias{try_parsing}
\title{try running an expression and capture either the output or error message.}
\usage{
try_parsing(x, error_message)
}
\arguments{
\item{x}{an expression in character format}

\item{error_message}{the message to return if the execution throws an error}
}
\value{
either the output of the execution or a message of class error_message
}
\description{
try running an expression and capture either the output or error message.
}
\examples{
 out <- try_parsing('1,w',error_message="SYTHAX ERROR")
 print(out)
 print(is_error(out))
}
