% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_running.R
\name{try_running}
\alias{try_running}
\title{try running an expression and capture either the output or error message.}
\usage{
try_running(x, message)
}
\arguments{
\item{x}{the expression}

\item{error}{the message to return if the execution throws an error}
}
\value{
either the output of the execution or a message of class error
}
\description{
try running an expression and capture either the output or error message.
}
\examples{
 out <- try_running(1+a_nonexistent_variable,message="Execution Error")
 print(out)
}
