% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave}
\alias{weave}
\title{weave two vectors or lists.}
\usage{
weave(x, y)
}
\arguments{
\item{x}{\code{[[vector|list]}}

\item{y}{\code{[class(x)]} must have the same class as \code{x} and have length equal to \code{[1 | length(x) | length(x)-1 ]}.  If length is 1, \code{y} will be inserted between every instance of \code{x}.}
}
\value{
\code{weave}: \code{[vector|list]}
}
\description{
weave two vectors or lists.
}
\examples{
 weave(letters, LETTERS)
 weave(letters, as.character(1:26))
 weave(letters, '.')
 weave(1:5, 6:9)
 weave(list(1, factor('G'), TRUE), list('k', 'k', 'l'))
}
