% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_splitn.R
\name{str_splitn}
\alias{str_splitn}
\title{Splits a string by line.}
\usage{
str_splitn(x, trim = TRUE, unlist = TRUE)
}
\arguments{
\item{x}{\code{[character]}}

\item{trim}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{unlist}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{str_splitn}: [character]
}
\description{
Splits a string by line.
}
\examples{
 str_splitn('#' NOTE: If you manually change your package name in the DESCRIPTION,\n  #' don't forget to change it here too, and in the config file.\n  #' For a safer name change mechanism, use the function.',
 trim = TRUE, unlist = TRUE)
 str_splitn('#' NOTE: If you manually change your package name in the DESCRIPTION,\n  #' don't forget to change it here too, and in the config file.\n  #' For a safer name change mechanism, use the function.',
 trim = FALSE, unlist = TRUE)
 str_splitn('#' NOTE: If you manually change your package name in the DESCRIPTION,\n  #' don't forget to change it here too, and in the config file.\n  #' For a safer name change mechanism, use the function.',
 trim = FALSE, unlist = FALSE)
}
