% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weave.R
\name{weave}
\alias{weave}
\title{weave to vectors or lists.}
\usage{
weave(x, y)
}
\value{
\code{weave}: \code{[vector|list]}
}
\description{
weave to vectors or lists.
}
\examples{
 weave(letters, LETTERS)
 weave(letters, as.character(1:26))
 weave(letters, '.')
 weave(1:5, 6:9)
 weave(list(1, factor('G'), TRUE), list('k', 'k', 'l'))
}
